moved button to cover-buttons component
