+++ src/components/landing/pages/intro/intro.tsx
@@ -1,17 +1,13 @@
 import React from 'react'
 import screenshot from './img/screenshot.png';
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
-import {Button} from 'react-bootstrap';
 import {Trans, useTranslation} from "react-i18next";
-import {Link} from "react-router-dom";
-import {useSelector} from "react-redux";
-import {ApplicationState} from "../../../../redux";
 import {FeatureLinks} from "./feature-links";
-import {LoginStatus} from "../../../../redux/user/types";
+import {CoverButtons} from "./cover-buttons";
 
 const Intro: React.FC = () => {
     useTranslation();
-    const user = useSelector((state: ApplicationState) => state.user);
+
 
     return (
         <div>
@@ -22,37 +18,7 @@ const Intro: React.FC = () => {
                 <Trans i18nKey="coverSlogan"/>
             </p>
 
-            {
-                user.status === LoginStatus.forbidden ?
-                    <div className="mb-5">
-                        <Link to="/login">
-                            <Button
-                                variant="success"
-                                size="lg"
-                                style={{minWidth: "200px"}}
-                            >
-                                <Trans i18nKey="signIn"/>
-                            </Button>
-                        </Link>
-
-
-                        <span className="m-2">
-                            <Trans i18nKey="or"/>
-                        </span>
-
-                        <Link to="/features">
-                            <Button
-                                variant="primary"
-                                size="lg"
-                                style={{minWidth: "200px"}}
-                            >
-                                <Trans i18nKey="exploreFeatures"/>
-                            </Button>
-                        </Link>
-                    </div>
-                :
-                    null
-            }
+            <CoverButtons/>
 
             <img alt="CodiMD Screenshot" src={screenshot} className="img-fluid mb-5"/>
             <FeatureLinks/>
